+++ cli/src/commands/fix.rs
@@ -13,11 +13,13 @@
 // limitations under the License.
 
 use std::collections::{HashMap, HashSet};
+use std::io::Write;
+use std::process::Stdio;
+use std::sync::mpsc::channel;
 
 use futures::StreamExt;
 use itertools::Itertools;
 use jj_lib::backend::{BackendError, BackendResult, CommitId, FileId, TreeValue};
-use jj_lib::commit::{Commit, CommitIteratorExt};
 use jj_lib::matchers::EverythingMatcher;
 use jj_lib::merged_tree::MergedTreeBuilder;
 use jj_lib::repo::Repo;
@@ -25,19 +27,48 @@ use jj_lib::repo_path::RepoPathBuf;
 use jj_lib::revset::{RevsetExpression, RevsetIteratorExt};
 use jj_lib::store::Store;
 use pollster::FutureExt;
+use rayon::iter::IntoParallelIterator;
+use rayon::prelude::ParallelIterator;
 use tracing::instrument;
 
 use crate::cli_util::{CommandHelper, RevisionArg};
-use crate::command_error::CommandError;
+use crate::command_error::{config_error_with_message, CommandError};
+use crate::config::CommandNameAndArgs;
 use crate::ui::Ui;
 
-/// Format files
+/// Update files with formatting fixes or other changes
+///
+/// The primary use case for this command is to apply the results of automatic
+/// code formatting tools to revisions that may not be properly formatted yet.
+/// It can also be used to modify files with other tools like `sed` or `sort`.
+///
+/// The changed files in the given revisions will be updated with any fixes
+/// determined by passing their file content through the external tool.
+/// Descendants will also be updated by passing their versions of the same files
+/// through the same external tool, which will never result in new conflicts.
+/// Files with existing conflicts will be updated on all sides of the conflict,
+/// which can potentially increase or decrease the number of conflict markers.
+///
+/// The external tool must accept the current file content on standard input,
+/// and return the updated file content on standard output. The output will not
+/// be used unless the tool exits with a successful exit code. Output on
+/// standard error will be passed through to the terminal.
+///
+/// The configuration schema is expected to change in the future. For now, it
+/// defines a single command that will affect all changed files in the specified
+/// revisions. For example, to format some Rust code changed in the working copy
+/// revision, you could write this configuration:
+///
+/// [fix]
+/// tool-command = ["rustfmt", "--emit", "stdout"]
+///
+/// And then run the command `jj fix -s @`.
 #[derive(clap::Args, Clone, Debug)]
 #[command(verbatim_doc_comment)]
 pub(crate) struct FixArgs {
-    /// Fix these commits and all their descendants
+    /// Fix files in the specified revision(s) and their descendants
     #[arg(long, short)]
-    sources: Vec<RevisionArg>,
+    source: Vec<RevisionArg>,
 }
 
 #[instrument(skip_all)]
@@ -47,107 +78,245 @@ pub(crate) fn cmd_fix(
     args: &FixArgs,
 ) -> Result<(), CommandError> {
     let mut workspace_command = command.workspace_helper(ui)?;
-    let root_commits: Vec<Commit> = workspace_command
-        .parse_union_revsets(&args.sources)?
-        .evaluate_to_commits()?
-        .try_collect()?;
-    workspace_command.check_rewritable(root_commits.iter().ids())?;
+    let root_commits: Vec<CommitId> = workspace_command
+        .parse_union_revsets(if args.source.is_empty() {
+            &[RevisionArg::AT]
+        } else {
+            &args.source
+        })?
+        .evaluate_to_commit_ids()?
+        .collect();
+    workspace_command.check_rewritable(root_commits.iter())?;
 
     let mut tx = workspace_command.start_transaction();
 
-    // Collect all FileIds we're going to format and which commits they appear in
-    let commits: Vec<_> = RevsetExpression::commits(root_commits.iter().ids().cloned().collect())
+    // Collect all of the unique `ToolInput`s we're going to use. Tools should be
+    // deterministic, and should not consider outside information, so it is safe to
+    // deduplicate inputs that correspond to multiple files or commits. This is
+    // typically more efficient, but it does prevent certain use cases like
+    // providing commit IDs as inputs to be inserted into files. We also need to
+    // record the mapping between tool inputs and paths/commits, to efficiently
+    // rewrite the commits later.
+    //
+    // If a path is being fixed in a particular commit, it must also be fixed in all
+    // that commit's descendants. We do this as a way of propagating changes,
+    // under the assumption that it is more useful than performing a rebase and
+    // risking merge conflicts. In the case of code formatters, rebasing wouldn't
+    // reliably produce well formatted code anyway. Deduplicating inputs helps
+    // to prevent quadratic growth in the number of tool executions required for
+    // doing this in long chains of commits with disjoint sets of modified files.
+    let commits: Vec<_> = RevsetExpression::commits(root_commits.to_vec())
         .descendants()
         .evaluate_programmatic(tx.base_repo().as_ref())?
         .iter()
         .commits(tx.repo().store())
         .try_collect()?;
-    let mut file_ids = HashSet::new();
-    let mut commit_paths: HashMap<CommitId, Vec<RepoPathBuf>> = HashMap::new();
+    let mut unique_tool_inputs: HashSet<ToolInput> = HashSet::new();
+    let mut commit_paths: HashMap<CommitId, HashSet<RepoPathBuf>> = HashMap::new();
     for commit in commits.iter().rev() {
-        let mut paths = vec![];
-        // Paths modified in parent commits in the set should also be updated in this
-        // commit
+        let mut paths: HashSet<RepoPathBuf> = HashSet::new();
+
+        // Fix all paths that were fixed in ancestors, so we don't lose those changes.
+        // We do this instead of rebasing onto those changes, to avoid merge conflicts.
         for parent_id in commit.parent_ids() {
             if let Some(parent_paths) = commit_paths.get(parent_id) {
-                paths.extend_from_slice(parent_paths);
+                paths.extend(parent_paths.iter().cloned());
             }
         }
-        let parent_tree = commit.parent_tree(tx.repo())?;
+
+        // Also fix any new paths that were changed in this commit.
         let tree = commit.tree()?;
+        let parent_tree = commit.parent_tree(tx.repo())?;
         let mut diff_stream = parent_tree.diff_stream(&tree, &EverythingMatcher);
         async {
             while let Some((repo_path, diff)) = diff_stream.next().await {
                 let (_before, after) = diff?;
+                // Deleted files have no file content to fix, and they have no terms in `after`,
+                // so we don't add any tool inputs for them. Conflicted files produce one tool
+                // input for each side of the conflict.
                 for term in after.into_iter().flatten() {
+                    // We currently only support fixing the content of normal files, so we skip
+                    // directories and symlinks, and we ignore the executable bit.
                     if let TreeValue::File { id, executable: _ } = term {
-                        file_ids.insert((repo_path.clone(), id));
-                        paths.push(repo_path.clone());
+                        // TODO: Consider filename arguments and tool configuration instead of
+                        // passing every changed file into the tool. Otherwise, the tool has to
+                        // be modified to implement that kind of stuff.
+                        let tool_input = ToolInput {
+                            file_id: id.clone(),
+                            repo_path: repo_path.clone(),
+                        };
+                        unique_tool_inputs.insert(tool_input.clone());
+                        paths.insert(repo_path.clone());
                     }
                 }
             }
             Ok::<(), BackendError>(())
         }
         .block_on()?;
+
         commit_paths.insert(commit.id().clone(), paths);
     }
 
-    let formatted = format_files(tx.repo().store().as_ref(), &file_ids)?;
+    // Run the configured tool on all of the chosen inputs.
+    // TODO: Support configuration of multiple tools and which files they affect.
+    let tool_command: CommandNameAndArgs = command
+        .settings()
+        .config()
+        .get("fix.tool-command")
+        .map_err(|err| config_error_with_message("Invalid `fix.tool-command`", err))?;
+    let fixed_file_ids = fix_file_ids(
+        tx.repo().store().as_ref(),
+        &tool_command,
+        &unique_tool_inputs,
+    )?;
 
+    // Substitute the fixed file IDs into all of the affected commits. Currently,
+    // fixes cannot delete or rename files, change the executable bit, or modify
+    // other parts of the commit like the description.
+    let mut num_checked_commits = 0;
+    let mut num_fixed_commits = 0;
     tx.mut_repo().transform_descendants(
         command.settings(),
-        root_commits.iter().ids().cloned().collect_vec(),
+        root_commits.iter().cloned().collect_vec(),
         |mut rewriter| {
-            let paths = commit_paths.get(rewriter.old_commit().id()).unwrap();
+            // TODO: Build the trees in parallel before `transform_descendants()` and only
+            // keep the tree IDs in memory, so we can pass them to the rewriter.
+            let repo_paths = commit_paths.get(rewriter.old_commit().id()).unwrap();
             let old_tree = rewriter.old_commit().tree()?;
             let mut tree_builder = MergedTreeBuilder::new(old_tree.id().clone());
-            for path in paths {
-                let old_value = old_tree.path_value(path);
+            let mut changes = 0;
+            for repo_path in repo_paths {
+                let old_value = old_tree.path_value(repo_path)?;
                 let new_value = old_value.map(|old_term| {
                     if let Some(TreeValue::File { id, executable }) = old_term {
-                        if let Some(new_id) = formatted.get(&(path, id)) {
-                            Some(TreeValue::File {
+                        let tool_input = ToolInput {
+                            file_id: id.clone(),
+                            repo_path: repo_path.clone(),
+                        };
+                        if let Some(new_id) = fixed_file_ids.get(&tool_input) {
+                            return Some(TreeValue::File {
                                 id: new_id.clone(),
                                 executable: *executable,
-                            })
-                        } else {
-                            old_term.clone()
+                            });
                         }
-                    } else {
-                        old_term.clone()
                     }
+                    old_term.clone()
                 });
                 if new_value != old_value {
-                    tree_builder.set_or_remove(path.clone(), new_value);
+                    tree_builder.set_or_remove(repo_path.clone(), new_value);
+                    changes += 1;
                 }
             }
-            let new_tree = tree_builder.write_tree(rewriter.mut_repo().store())?;
-            let builder = rewriter.reparent(command.settings())?;
-            builder.set_tree_id(new_tree).write()?;
+            num_checked_commits += 1;
+            if changes > 0 {
+                num_fixed_commits += 1;
+                let new_tree = tree_builder.write_tree(rewriter.mut_repo().store())?;
+                let builder = rewriter.reparent(command.settings())?;
+                builder.set_tree_id(new_tree).write()?;
+            }
             Ok(())
         },
     )?;
-
-    tx.finish(ui, format!("fixed {} commits", root_commits.len()))
+    writeln!(
+        ui.status(),
+        "Fixed {num_fixed_commits} commits of {num_checked_commits} checked."
+    )?;
+    tx.finish(ui, format!("fixed {num_fixed_commits} commits"))
 }
 
-fn format_files<'a>(
+/// Represents the API between `jj fix` and the tools it runs.
+// TODO: Add the set of changed line/byte ranges, so those can be passed into code formatters via
+// flags. This will help avoid introducing unrelated changes when working on code with out of date
+// formatting.
+#[derive(PartialEq, Eq, Hash, Clone)]
+struct ToolInput {
+    /// File content is the primary input, provided on the tool's standard
+    /// input. We use the `FileId` as a placeholder here, so we can hold all
+    /// the inputs in memory without also holding all the content at once.
+    file_id: FileId,
+
+    /// The path is provided to allow passing it into the tool so it can
+    /// potentially:
+    ///  - Choose different behaviors for different file names, extensions, etc.
+    ///  - Update parts of the file's content that should be derived from the
+    ///    file's path.
+    repo_path: RepoPathBuf,
+}
+
+/// Applies `run_tool()` to the inputs and stores the resulting file content.
+///
+/// Returns a map describing the subset of `tool_inputs` that resulted in
+/// changed file content. Failures when handling an input will cause it to be
+/// omitted from the return value, which is indistinguishable from succeeding
+/// with no changes.
+/// TODO: Better error handling so we can tell the user what went wrong with
+/// each failed input.
+fn fix_file_ids<'a>(
     store: &Store,
-    file_ids: &'a HashSet<(RepoPathBuf, FileId)>,
-) -> BackendResult<HashMap<(&'a RepoPathBuf, &'a FileId), FileId>> {
+    tool_command: &CommandNameAndArgs,
+    tool_inputs: &'a HashSet<ToolInput>,
+) -> BackendResult<HashMap<&'a ToolInput, FileId>> {
+    let (updates_tx, updates_rx) = channel();
+    // TODO: Switch to futures, or document the decision not to. We don't need
+    // threads unless the threads will be doing more than waiting for pipes.
+    tool_inputs.into_par_iter().try_for_each_init(
+        || updates_tx.clone(),
+        |updates_tx, tool_input| -> Result<(), BackendError> {
+            let mut read = store.read_file(&tool_input.repo_path, &tool_input.file_id)?;
+            let mut old_content = vec![];
+            read.read_to_end(&mut old_content).unwrap();
+            if let Ok(new_content) = run_tool(tool_command, tool_input, &old_content) {
+                if new_content != *old_content {
+                    let new_file_id =
+                        store.write_file(&tool_input.repo_path, &mut new_content.as_slice())?;
+                    updates_tx.send((tool_input, new_file_id)).unwrap();
+                }
+            }
+            Ok(())
+        },
+    )?;
+    drop(updates_tx);
     let mut result = HashMap::new();
-    for (path, id) in file_ids {
-        // TODO: read asynchronously
-        let mut read = store.read_file(path, id)?;
-        let mut buf = vec![];
-        read.read_to_end(&mut buf).unwrap();
-        // TODO: Call a formatter instead of just uppercasing
-        for b in &mut buf {
-            b.make_ascii_uppercase();
-        }
-        // TODO: Don't write it if it didn't change
-        let formatted_id = store.write_file(path, &mut buf.as_slice())?;
-        result.insert((path, id), formatted_id);
+    while let Ok((tool_input, new_file_id)) = updates_rx.recv() {
+        result.insert(tool_input, new_file_id);
     }
     Ok(result)
+}
+
+/// Runs the `tool_command` to fix the given file content.
+///
+/// The `old_content` is assumed to be that of the `tool_input`'s `FileId`, but
+/// this is not verified.
+///
+/// Returns the new file content, whose value will be the same as `old_content`
+/// unless the command introduced changes. Returns `None` if there were any
+/// failures when starting, stopping, or communicating with the subprocess.
+fn run_tool(
+    tool_command: &CommandNameAndArgs,
+    tool_input: &ToolInput,
+    old_content: &[u8],
+) -> Result<Vec<u8>, ()> {
+    // TODO: Pipe stderr so we can tell the user which commit, file, and tool it is
+    // associated with.
+    let mut vars: HashMap<&str, &str> = HashMap::new();
+    vars.insert("path", tool_input.repo_path.as_internal_file_string());
+    let mut child = tool_command
+        .to_command_with_variables(&vars)
+        .stdin(Stdio::piped())
+        .stdout(Stdio::piped())
+        .spawn()
+        .or(Err(()))?;
+    let mut stdin = child.stdin.take().unwrap();
+    let output = std::thread::scope(|s| {
+        s.spawn(move || {
+            stdin.write_all(old_content).ok();
+        });
+        Some(child.wait_with_output().or(Err(())))
+    })
+    .unwrap()?;
+    if output.status.success() {
+        Ok(output.stdout)
+    } else {
+        Err(())
+    }
 }